Split stat related APIs into two files by point/series
