+++ server/stat_test.go
